Added Environmental Inspection Form and enhanced UI
An Environmental Inspection Form has been introduced to the Tab Navigation. This allows users to create new environmental inspections directly from the Health Safety navigator. In addition, some changes were made to enhance the readability of form sections, including increasing the text size in the CustomHsHeader, adjusting the font style, and modifying the padding for a more balanced look. The modifications are expected to make the interface more user-friendly and efficient.

Issue References:
Related to issue #254