Add SVG Icon support with touch functionality
This commit introduces a new, reusable SVG Icon component to the project that is both modern, efficient and lightweight. This SVG Icon component supports customizable size and color and, more importantly, the component is tappable with an optional press handling function. Consequently, this introduces more versatility, increasing the app's potential for interactions.

Moreover, the icon set has been added to the Utils directory and the first icon, 'deleteIcon', has been defined as proof of concept.

Future icon additions simply require extending this icon set. This offers a simplified and centralized manageability for all icons in the project. This change continues our transition towards a more modern, efficient, and scalable codebase.

Finally, a hitSlop prop for adjusting the touchable area around the Icon for better user experience has been introduced.