Optimize image processing and improve form data management
Many changes to the logic of saving and processing images were made. Images are now optimized before they are saved - they are resized and compressed, resulting in reduced resolution and smaller file sizes. Furthermore, images are now processed in limited batches instead of individually, to increase performance and manage resources more effectively.

Additionally, the way form data is managed in relation to section changes in both DcFormCreate.jsx and EnvCreateForm.jsx has been significantly improved. Now, data related to the current form section is explicitly saved prior to a section change, and the existing data for the new section (if any) is loaded. It makes the form's data management approach more robust and intuitive.

Lastly, an enhancement has been made in the image saving process to ensure each save operation completes successfully. Now, if an error is encountered, an alert is displayed, and the process is interrupted, avoiding saving incorrect data.

No associated issues.